Check for files status in the Image and LabeledImage tables
Added to the tables a status column. When the Image and LabeledImage row is
saved, the model will try to load the file and saves the results in the
filestate field.

Also added a web command `check_filenames` that will check all filename field of
Image and LabeledImage tables.

Fixes #8